DeleteArtist_404_url_for
When attempting to hit the delete button next to our artists to monitor, we get the error 404 message citing 'delete_artist' html page is not found. It is intended to redirect to the '/delete_artist' route, where the function deletes the artist record from the DB & then redirects back to the scrapelist2 page, with an updated list reflecting our deletion.